refactor: moving the jobPath setting from JobPath to JobDB
