+++ src/DIRAC/WorkloadManagementSystem/ConfigTemplate.cfg
@@ -361,11 +361,7 @@ Executors
 {
   Optimizers
   {
-    Load = JobPath, JobSanity, InputData, JobScheduling, CheckingToStagingTransitioner, CheckingToWaitingTransitioner
-  }
-  JobPath
-  {
-
+    Load = JobSanity, InputData, JobScheduling, CheckingToStagingTransitioner, CheckingToWaitingTransitioner
   }
   JobSanity
   {
