+++ src/DIRAC/WorkloadManagementSystem/DB/JobDB.py
@@ -12,21 +12,17 @@ The following options can be set in ``Systems/WorkloadManagement/<Setup>/Databas
 
 """
 import base64
-import re
 import zlib
 import datetime
 
 import operator
 
 from DIRAC import S_OK, S_ERROR
-from DIRAC.ConfigurationSystem.Client.Helpers.Registry import getVOForGroup
-from DIRAC.ConfigurationSystem.Client.Helpers.Operations import Operations
 from DIRAC.ConfigurationSystem.Client.Helpers.Resources import getSiteTier
 from DIRAC.Core.Base.DB import DB
-from DIRAC.Core.Utilities import DErrno, List
+from DIRAC.Core.Utilities import DErrno
 from DIRAC.Core.Utilities.ClassAd import ClassAd
 from DIRAC.Core.Utilities.DErrno import EWMSSUBM, EWMSJMAN
-from DIRAC.Core.Utilities.ObjectLoader import ObjectLoader
 from DIRAC.ResourceStatusSystem.Client.SiteStatus import SiteStatus
 from DIRAC.WorkloadManagementSystem.Client import JobStatus
 from DIRAC.WorkloadManagementSystem.Client import JobMinorStatus
@@ -920,16 +916,24 @@ class JobDB(DB):
         :return: new job ID
         """
 
+        # 1.- insert original JDL on DB and get new JobID
+        result = self.__insertNewJDL(jdl)
+        if not result["OK"]:
+            return S_ERROR(EWMSSUBM, "Failed to insert JDL in to DB")
+        jobID = result["Value"]
+
+        # Replace the JobID placeholder if any
+        if jdl.find("%j") != -1:
+            jdl = jdl.replace("%j", str(jobID))
+
+        # Set the JobID attribute of the manifest
         try:
             jobDescription = ClassAd(jdl)
         except SyntaxError as e:
             return S_ERROR(e)
 
-        if jobDescription.isEmpty():
-            return S_ERROR(EWMSSUBM, "The jdl is not OK")
-
-        # Resolve
-        result = resolveJobDescription(jobDescription, owner, ownerDN, ownerGroup, diracSetup)
+        # Insert the job jdl with the JobID
+        result = self.setJobJDL(jobID, jobDescription.asJDL())
         if not result["OK"]:
             return result
         jobManifest.setOptionsFromDict({"Owner": owner, "OwnerDN": ownerDN, "OwnerGroup": ownerGroup})
@@ -990,19 +994,6 @@ class JobDB(DB):
 
         jobAttrNames.append("MinorStatus")
         jobAttrValues.append(initialMinorStatus)
-
-        # Set the JobID attribute of the manifest
-        jobDescription.insertAttributeInt("JobID", jobID)
-
-        jobJDL = jobDescription.asJDL()
-
-        # Replace the JobID placeholder if any
-        if jobJDL.find("%j") != -1:
-            jobJDL = jobJDL.replace("%j", str(jobID))
-
-        result = self.setJobJDL(jobID, jobJDL)
-        if not result["OK"]:
-            return result
 
         # Adding the job in the Jobs table
         result = self.insertFields("Jobs", jobAttrNames, jobAttrValues)
@@ -1182,33 +1173,7 @@ class JobDB(DB):
         if not self._update(f"DELETE FROM OptimizerParameters WHERE JobID={e_jobID}")["OK"]:
             return S_ERROR("JobDB.removeJobOptParameter: operation failed.")
 
-        # the JobManager needs to know if there is InputData ??? to decide which optimizer to call
-        # proposal: - use the getInputData method
-        res = self.getJobJDL(jobID, original=True)
-        if not res["OK"]:
-            return res
-
-        try:
-            classAdJob = ClassAd(res["Value"])
-        except SyntaxError as e:
-            return S_ERROR(e)
-
-        classAdJob.insertAttributeInt("JobID", jobID)
-
-        jobAttrNames.append("UserPriority")
-        jobAttrValues.append(classAdJob.getAttributeInt("Priority"))
-
-        siteList = classAdJob.getListFromExpression("Site")
-        if not siteList:
-            site = "ANY"
-        elif len(siteList) > 1:
-            site = "Multiple"
-        else:
-            site = siteList[0]
-
-        jobAttrNames.append("Site")
-        jobAttrValues.append(site)
-
+        # Update some attributes
         jobAttrNames.append("Status")
         jobAttrValues.append(JobStatus.RECEIVED)
 
@@ -1227,27 +1192,12 @@ class JobDB(DB):
         jobAttrNames.append("RescheduleTime")
         jobAttrValues.append(str(datetime.datetime.utcnow()))
 
-        jobJDL = classAdJob.asJDL()
-
-        # Replace the JobID placeholder if any
-        if jobJDL.find("%j") != -1:
-            jobJDL = jobJDL.replace("%j", str(jobID))
-
-        result = self.setJobJDL(jobID, jobJDL)
-        if not result["OK"]:
-            return result
-
-        result = self.__setInitialJobParameters(classAdJob, jobID)
-        if not result["OK"]:
-            return result
-
         result = self.setJobAttributes(jobID, jobAttrNames, jobAttrValues, force=True)
         if not result["OK"]:
             return result
 
         retVal = S_OK(jobID)
         retVal["JobID"] = jobID
-        retVal["InputData"] = classAdJob.lookupAttribute("InputData")
         retVal["RescheduleCounter"] = rescheduleCounter
         retVal["Status"] = JobStatus.RECEIVED
         retVal["MinorStatus"] = JobMinorStatus.RESCHEDULED
@@ -1904,201 +1854,3 @@ class JobDB(DB):
         result = self._update(cmd)
         self.log.verbose("Removed from HBLI", result)
         return result
-
-
-def resolveJobDescription(jobDescription: ClassAd, owner: str, ownerDN: str, ownerGroup: str, diracSetup: str):
-    """Method that launch all the job description resolvers"""
-
-    opsHelper = Operations(group=ownerGroup, setup=diracSetup)
-
-    # Resolve owner
-    jobDescription.insertAttributeString("DIRACSetup", diracSetup)
-    jobDescription.insertAttributeString("Owner", owner)
-    jobDescription.insertAttributeString("OwnerDN", ownerDN)
-    jobDescription.insertAttributeString("OwnerGroup", ownerGroup)
-
-    # Resolve VO
-    vo = getVOForGroup(ownerGroup)
-    if vo:
-        jobDescription.insertAttributeString("VirtualOrganization", vo)
-
-    # Resolve single elements into a list
-    resolveSingularNamesToPlurals(jobDescription)
-
-    # Resolve Input data module
-    inputDataPolicy = Operations(vo=vo).getValue("InputDataPolicy/InputDataModule")
-    if inputDataPolicy and not jobDescription.lookupAttribute("InputDataModule"):
-        jobDescription.insertAttributeString("InputDataModule", inputDataPolicy)
-
-    # Resolve priority
-    resolvePriority(jobDescription, opsHelper)
-
-    # Resolve CPU time
-    resolveCpuTime(jobDescription, opsHelper)
-
-    # Resolve sites
-    resolveSites(jobDescription)
-
-    # Resolve tags
-    resolveTags(jobDescription)
-
-    return S_OK()
-
-
-def resolveSingularNamesToPlurals(jobDescription: ClassAd) -> None:
-    """Resolve the old singular naming convention to plurals"""
-    translationalDictionary = {"BannedSite": "BannedSites", "GridCE": "GridCEs", "Site": "Sites", "Tag": "Tags"}
-    for oldName, newName in translationalDictionary.items():
-        if jobDescription.lookupAttribute(oldName):
-            jobDescription.insertAttributeVectorString(
-                newName, List.fromChar(jobDescription.getAttributeString(oldName))
-            )
-            # TODO: replace old name by new names everywhere in the code
-            # so we can delete the old attributes from the job description
-            # jobDescription.deleteAttribute(oldName)
-
-
-def resolveTags(jobDescription: ClassAd) -> None:
-    """Resolve a list of tagsTQ from the job manifest content for the TQ"""
-
-    tags = set()
-
-    # CPU cores
-    if jobDescription.lookupAttribute("NumberOfProcessors"):
-        minProcessors = maxProcessors = jobDescription.getAttributeInt("NumberOfProcessors")
-    else:
-        if jobDescription.lookupAttribute("MinNumberOfProcessors"):
-            minProcessors = jobDescription.getAttributeInt("MinNumberOfProcessors")
-        else:
-            minProcessors = 1
-        if jobDescription.lookupAttribute("MaxNumberOfProcessors"):
-            maxProcessors = jobDescription.getAttributeInt("MaxNumberOfProcessors")
-        else:
-            maxProcessors = minProcessors
-
-    if minProcessors > 1:
-        tags.add(f"{minProcessors}Processors")
-    if maxProcessors > 1:
-        tags.add("MultiProcessor")
-
-    # Whole node
-    if jobDescription.lookupAttribute("WholeNode"):
-        if jobDescription.getAttributeString("WholeNode").lower() in ["1", "yes", "true", "y"]:
-            tags.add("WholeNode")
-            tags.add("MultiProcessor")
-
-    # RAM
-    if jobDescription.lookupAttribute("MaxRAM"):
-        maxRAM = jobDescription.getAttributeInt("MaxRAM")
-        if maxRAM:
-            tags.add(f"{maxRAM}GB")
-
-    # Other tags? Just add them
-    if jobDescription.lookupAttribute("Tags"):
-        tags |= set(jobDescription.getListFromExpression("Tags"))
-
-    # Store in the job description the tags if any
-    if tags:
-        jobDescription.insertAttributeVectorString("Tags", tags)
-
-
-def resolvePriority(jobDescription: ClassAd, operations: Operations) -> None:
-    """Resolve the job priority and stores it in the job description"""
-    if jobDescription.lookupAttribute("Priority"):
-        minPriority = operations.getValue("JobDescription/MinPriority", 0)
-        maxPriority = operations.getValue("JobDescription/MaxPriority", 10)
-        priority = max(minPriority, min(jobDescription.getAttributeInt("Priority"), maxPriority))
-    else:
-        priority = operations.getValue("JobDescription/DefaultPriority", 1)
-
-    jobDescription.insertAttributeInt("Priority", int(priority))
-
-
-def resolveCpuTime(jobDescription: ClassAd, operations: Operations) -> None:
-    """Resolve the CPU time and stores it in the job description"""
-    if jobDescription.lookupAttribute("CPUTime"):
-        minCpuTime = operations.getValue("JobDescription/MinCPUTime", 100)
-        maxCpuTime = operations.getValue("JobDescription/MaxCPUTime", 500000)
-        resolvedCpuTime = max(minCpuTime, min(jobDescription.getAttributeInt("CPUTime"), maxCpuTime))
-    else:
-        resolvedCpuTime = operations.getValue("JobDescription/DefaultCPUTime", 86400)
-
-    jobDescription.insertAttributeInt("CPUTime", int(resolvedCpuTime))
-
-
-def resolveSites(jobDescription: ClassAd):
-    """Remove the Site attribute if set to ANY in the original manifest"""
-
-    if jobDescription.lookupAttribute("Sites"):
-        sites = jobDescription.getListFromExpression("Sites")
-        if not sites or "ANY" in sites or "Any" in sites or "any" in sites:
-            jobDescription.deleteAttribute("Sites")
-
-def checkJobDescription(jobDescription: ClassAd):
-    """Check that the job description is correctly set"""
-
-    # Check that the job description contains some mandatory fields
-    mandatoryFields = {"JobType", "Owner", "OwnerDN", "OwnerGroup", "DIRACSetup", "Priority"}
-    for field in mandatoryFields:
-        if not jobDescription.lookupAttribute(field):
-            return S_ERROR(EWMSSUBM, f"The JDL needs to contain the {field} field")
-
-    # Check that job type is correct
-    opsHelper = Operations(
-        group=jobDescription.getAttributeString("OwnerGroup"), setup=jobDescription.getAttributeString("DIRACSetup")
-    )
-    allowedJobTypes = opsHelper.getValue("JobDescription/AllowedJobTypes", ["User", "Test", "Hospital"])
-    transformationTypes = opsHelper.getValue("Transformations/DataProcessing", [])
-    jobTypes = opsHelper.getValue("JobDescription/ChoicesJobType", allowedJobTypes + transformationTypes)
-    jobType = jobDescription.getAttributeString("JobType")
-    if jobType not in jobTypes:
-        return S_ERROR(EWMSSUBM, f"{jobType} is not a valid value for JobType")
-
-    # Check input data
-    inputData = jobDescription.getListFromExpression("InputData")
-    if inputData:
-
-        # Check that the VO is set up
-        vo = jobDescription.getListFromExpression("VirtualOrganization")
-        if not vo:
-            return S_ERROR(EWMSSUBM, "Input data listed but no VO is set")
-
-        # Check that LFNs are well formated
-        voRE = re.compile(f"^(LFN:)?/{vo}/")
-        for lfn in inputData:
-            if not voRE.match(lfn):
-                return S_ERROR(EWMSSUBM, JobMinorStatus.INPUT_INCORRECT)
-            if lfn.find("//") > -1:
-                return S_ERROR(EWMSSUBM, JobMinorStatus.INPUT_CONTAINS_SLASHES)
-
-        # Check max input data
-        maxInputData = Operations().getValue("JobDescription/MaxInputData", 500)
-        if len(inputData) > maxInputData:
-            return S_ERROR(
-                EWMSSUBM,
-                f"Number of Input Data Files ({len(inputData)}) greater than current limit: {maxInputData}",
-            )
-
-    # Check LFN input sandboxes
-    inputSandboxes = jobDescription.getListFromExpression("InputSandbox")
-    for inputSandbox in inputSandboxes:
-        if inputSandbox.startswith("LFN:") and not inputSandbox.startswith("LFN:/"):
-            return S_ERROR(EWMSSUBM, "LFNs should always start with '/'")
-
-    # Check JobPath
-    jobPath = jobDescription.getListFromExpression("JobPath")
-    for optimizer in jobPath:
-        result = ObjectLoader().loadObject(f"WorkloadManagementSystem.Executors.{optimizer}")
-        if not result["OK"]:
-            return result
-
-    return S_OK()
-
-
-def getDIRACPlatform(platformList):
-    """Loading the function that will be used to determine the platform (it can be VO specific)"""
-    result = ObjectLoader().loadObject("ConfigurationSystem.Client.Helpers.Resources", "getDIRACPlatform")
-    if not result["OK"]:
-        return result
-    getDIRACPlatformMethod = result["Value"]
-    return getDIRACPlatformMethod(platformList)
